Expand ulox samples with inner Vms
